pyproject.toml
[tool.poetry]
name = "plus-planner-core"
version = "0.1.0"
description = ""
authors = ["Pedro Santana"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.1"
fastapi-pagination = "^0.12.22"
python-dotenv = "^1.0.1"
invoke = "^2.2.0"
uvicorn = "^0.29.0"
asyncpg = "^0.29.0"
python-multipart = "^0.0.9"
pydantic = "^2.6.4"
bcrypt = "^4.1.2"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
pyyaml = "5.3.1"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
fastapi-filter = "^1.1.0"
email-validator = "^2.1.1"
tortoise-orm = "^0.21.2"
plus_db_agent = { git = "https://github.com/pedrogs97/plus_db_agent.git", branch = "main" }
requests = "^2.32.3"
websockets = "^12.0"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
